avances persistencia

+++ src/main/java/sge/dispositivo/EstandarToInteligenteAdapter.java
@@ -2,7 +2,7 @@ package sge.dispositivo;
 
 public class EstandarToInteligenteAdapter extends DispositivoInteligente {
 
-	public EstandarToInteligenteAdapter(int idDispositivo, String nombreDispositivo, double consumoKwH,
+	public EstandarToInteligenteAdapter(Long idDispositivo, String nombreDispositivo, double consumoKwH,
 			String tipoDispositivo) {
 		
 		this.setIdDispositivo(idDispositivo);
